Moving utility to LikesUtilsProvider.
